updated the first plot with all of the initial traits (no new ones)
had to change notation in first line, Tash removed the captial letters.

No longer need dataframe for just trophic level, I figured out how to make that work like the rest of the columns.

0/1 columns wouldn't pipe properly as integers so I switched them to factors.

No longer need the weird offset that we did before (where only half of the heatmap box showed up). So I fixed the offset.

Removed 'benthopelagic from vert_habitat legend, as it is no longer in the dataframe

I found a better way to pipe the columns onto the tree. Before we had to call the column we wanted, as well as the species column, which we did by column number. I fixed this by putting 'my_prey_traits[ , "column", drop = FALSE]". With that calling method it works perfectly, as long as the species names are the rownames.

Removed "UN" category from diel migrant, refuge, and physical defense since that value is no longer in the data. Meaning the NA's are left as blanks, rather than having their own piece in the legend.

I commented out the part where we annotate the column name onto the plot itself. Now that we have more data, the names end up on top of the heatmap tiles. So I'm not sure of a solution to that right now, but I'll try to think of one